[refactor] moving memtracer to gemini (#790)
* [refactor] moving memtracer to gemini

* polish

* fix the stm bug

* move gemini unitest to an individual dir

* close tests/test_data_pipeline_tensor_parallel/test_cifar_with_data_pipeline_tensor.py

* remove imports in class

* remove some unitests

* revert pipeline unittest
