--- colossalai/utils/memory_tracer/__init__.py
@@ -1,4 +1,0 @@
-from .memory_monitor import AsyncMemoryMonitor, SyncCudaMemoryMonitor
-from .memstats_collector import MemStatsCollector
-
-__all__ = ['AsyncMemoryMonitor', 'SyncCudaMemoryMonitor', 'MemStatsCollector']
