+++ docs/colossalai/colossalai.utils.memory_tracer.memstats_collector.rst
@@ -1,5 +1,5 @@
 colossalai.utils.memory\_tracer.memstats\_collector
 ===================================================
 
-.. automodule:: colossalai.utils.memory_tracer.memstats_collector
+.. automodule:: colossalai.gemini.memory_tracer.memstats_collector
    :members:
